pyproject.toml
[project]
name = "the-automaton-auditor"
version = "0.1.0"
description = "Week 2 Digital Courtroom built with LangGraph"
requires-python = ">=3.10"
dependencies = [
    "langgraph",
    "langchain-core",
    "pydantic>=2.0.0",
    "typing_extensions>=4.8.0",
    "docling",
    "python-dotenv",
    "tenacity",
    "requests",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
]

[tool.uv]
index-url = "https://pypi.org/simple"

